Update issue templates
Updated the Bug report template to be more suitable for our project's needs: removed unnecessary smartphone data since this project is for a Desktop version, added the listing of the processor information, graphics information and JDK Version to be able to recreate the bug if needed. Also updated the Feature request template to be more descriptive so that users can clearly list the issue for more precise problem solving.